% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avian_document.R
\docType{data}
\name{avian}
\alias{avian}
\title{Avian treatment matrix}
\format{
A data frame with 319 rows and 1 variable:
\describe{
\item{individuals}{similarities, calcualted from PSI}
...
}
}
\source{
\url{https://github.com/Messer-Sealey-White-ADA2020/Vignette-Assignment/blob/master/blacktreatments_heads2020f.csv}
}
\usage{
avian
}
\description{
Antipredator behavior similarities for the avian treatment dataset
}
\keyword{datasets}
