% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltme.R
\name{igraphme}
\alias{igraphme}
\title{Convert meltme product to igraph format}
\usage{
igraphme(z)
}
\arguments{
\item{z}{igraph dataset to be converted into visNetwork}
}
\description{
This function converts a melted dataset (ideally from the meltme function) into an igraph framework
}
\examples{
igraphme(meltme_product)
}
\keyword{igraph,}
\keyword{melt}
