% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltme.R
\name{meltme}
\alias{meltme}
\title{A matrix melting function}
\usage{
meltme(x)
}
\arguments{
\item{x}{dataset to be melted}
}
\description{
This function melts a matrix into a format that can then be read in igraph or visNetwork for data visualization
}
\examples{
meltme(avian)
}
\keyword{matrix}
\keyword{melt,}
