% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltme_product.R
\docType{data}
\name{meltme_product}
\alias{meltme_product}
\title{Melted matrix of "avian" dataset}
\format{
A data frame with 406 rows and 4 variables:
\describe{
\item{ind1}{numeric form of individual A in comparison}
\item{ind2}{numeric form of individual B in comparison}
\item{values}{similarity between individual A & B antipredator display}
...
}
}
\source{
\url{https://raw.githubusercontent.com/Messer-Sealey-White-ADA2020/Vignette-Assignment/master/meltme_product.csv}
}
\usage{
meltme_product
}
\description{
The melted matrix of the avian dataset for demonstration of the rename_me
}
\keyword{datasets}
