% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monkey.R
\docType{data}
\name{monkey}
\alias{monkey}
\title{Monkey socialability matrix}
\format{
A data frame with 11 rows and 12 variables:
\describe{
\item{Anita}{individual in dataset}
\item{Carlos}{individual in dataset}
\item{Chico}{individual in dataset}
\item{Junon}{individual in dataset}
\item{Kato}{individual in dataset}
\item{Manuel}{individual in dataset}
\item{Penelope}{individual in dataset}
\item{Popeye}{individual in dataset}
\item{Reuben}{individual in dataset}
\item{Rosa}{individual in dataset}
\item{Sol}{individual in dataset}
\item{name}{labels individuals names}
...
}
}
\source{
\url{https://raw.githubusercontent.com/Messer-Sealey-White-ADA2020/Vignette-Assignment/master/EM_monkey.csv}
}
\usage{
monkey
}
\description{
Socialability indices for monkey dataset
}
\keyword{datasets}
